requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.2
Django==5.1.6
django-appconf==1.1.0
django-ckeditor-5==0.2.11
django-compressor==4.4
django-cors-headers==4.3.1
django-csp==3.8
django-extensions==3.2.3
django-photologue==3.15.1
django-ratelimit==4.1.0
django-redis==5.4.0
django-sortedm2m==4.0.0
ExifRead==3.0.0
google-api-python-client==2.123.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
gunicorn==21.2.0
httplib2==0.22.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
rcssmin==1.1.1
rjsmin==1.2.1
soupsieve==2.5
sqlparse==0.5.3
tzdata==2025.1
whitenoise==6.6.0
